+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 We follow [Semantic Versions](https://semver.org/).
 
 
-## Version 0.1.2 (unreleased)
+## Version 0.1.2
 
 - Add class `Post` - representing the Reddit post
 - Add class `Scheduler` - managing data about scheduled posts